printk: console: Implement core per-console loglevel infrastructure
Consoles can have vastly different latencies and throughputs. For
example, writing a message to the serial console can take on the order
of tens of milliseconds to get the UART to successfully write a message.
While this might be fine for a single, one-off message, this can cause
significant application-level stalls in situations where the kernel
writes large amounts of information to the console.

This means that while you might want to send at least INFO level
messages to (for example) netconsole, which is relatively fast, you may
only want to send at least WARN level messages to the serial console.
Such an implementation would permit debugging using the serial console
in cases that netconsole doesn't receive messages during particularly
bad system issues, while still keeping the noise low enough to avoid
inducing latency in userspace applications. To mitigate this, add such
an interface, extending the existing console loglevel controls to allow
each console to have its own loglevel.

One can't just disable the serial console, because one may actually need
it in situations where the machine is in a bad enough state that nothing
is received on netconsole. One also can't just bump the loglevel at
runtime after the issue, because usually the machine is already so
wedged by this point that it isn't responsive to such requests.

This commit adds the internal infrastructure to support per-console log
levels, which will be configurable through sysfs and the kernel command
line in future commits.

The global console_loglevel is preserved and used as the default log
level for all consoles. Each console can override this global level with
its own specific log level stored in its struct console. To override the
global level, the per-console log level must be greater than 0;
otherwise the default value of -1 makes it so that the global level is
used.

The existing ignore_loglevel command line parameter maintains behaviour
befitting its existing intent, overriding both the global and
per-console log levels.

Signed-off-by: Chris Down <chris@chrisdown.name>