printk: Constrain hardware-addressed console checks to name position
Commit 7640f1a44eba ("printk: Add
match_devname_and_update_preferred_console()") adds support for
DEVNAME:n:n style console= registration. It determines whether or not
this is a hardware-addressed console by looking at whether the console=
string has a colon in it. However, this interferes with loglevel:n and
potentially other future named options, which also make use of the
character.

In order to avoid this, only search positions before options.

Signed-off-by: Chris Down <chris@chrisdown.name>
Reviewed-by: Tony Lindgren <tony.lindgren@linux.intel.com>