printk: Override per-console loglevel on sysrq loglevel change
During emergency situations where one uses sysrq, it may not be feasible
for the operator to access sysfs or the sysctl interface. This would be
problematic in the case where the console the operator is looking at has
a per-console loglevel set, as sysrq loglevel setting would have no
effect.

Change the sysrq code to unset all per-console loglevels when doing
sysrq loglevel setting. This allows the global console loglevel to take
precedence. A warning is logged if any per-console loglevel is
overridden so that the user understands that this is a sustained effect.

Signed-off-by: Chris Down <chris@chrisdown.name>